LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-self-baseline-not-applied-if-sizing-cyclic-dependency-002-expected.txt

auto-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized rows - horizonal grid and verticalLR item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized rows - horizonal grid and verticalLR item - column-axis baseline
baseline is not applied initially, but orthogonal items force repeating the track sizing and height is not indefinite in that phase.
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 auto-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized columns - horizontal grid item - row-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized columns - horizontal grid item - row-axis baseline
baseline is not applied initially, but orthogonal items force repeating the track sizing and height is not indefinite in that phase.
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 auto-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized rows - verticalLR grid and horizontal item - column-axis baseline
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized rows - verticalLR grid and horizontal item - column-axis baseline
baseline is not applied initially, but orthogonal items force repeating the track sizing and height is not indefinite in that phase.
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 auto-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 min-content-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-content-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 fit-content-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 flex-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É
 max-flex-sized rows - verticalRL grid and horizontal item - column-axis baseline
baseline alignment is not supported, because the RL items do not share an alignment context with the LR items
É
ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É

FAIL .inline-grid 1 assert_equals:
<div class="inline-grid alignItemsBaseline columns height200">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="95" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="135">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="135" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 95 but got 350
FAIL .inline-grid 2 assert_equals:
<div class="inline-grid alignItemsBaseline columns min-content-rows">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="55" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="95">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="95" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 55 but got 350
FAIL .inline-grid 3 assert_equals:
<div class="inline-grid alignItemsBaseline columns max-content-rows">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="55" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="95">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="95" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 55 but got 350
FAIL .inline-grid 4 assert_equals:
<div class="inline-grid alignItemsBaseline columns fit-content-rows">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="55" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="95">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="95" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 55 but got 350
FAIL .inline-grid 5 assert_equals:
<div class="inline-grid alignItemsBaseline columns flex-rows">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="55" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="95">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="95" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 55 but got 350
FAIL .inline-grid 6 assert_equals:
<div class="inline-grid alignItemsBaseline columns auto-rows">
  <div class="firstRowFirstColumn bigFont paddingBottom" data-offset-x="0" data-offset-y="55" data-expected-width="100" data-expected-height="75">É</div>
  <div class="firstRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="95">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 height25" data-offset-x="0" data-offset-y="95" data-expected-width="200" data-expected-height="25"></div>
</div>
offsetTop expected 55 but got 350
PASS .inline-grid 7
PASS .inline-grid 8
PASS .inline-grid 9
PASS .inline-grid 10
PASS .inline-grid 11
PASS .inline-grid 12
PASS .inline-grid 13
PASS .inline-grid 14
PASS .inline-grid 15
PASS .inline-grid 16
PASS .inline-grid 17
PASS .inline-grid 18
FAIL .inline-grid 19 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline columns width200">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="30" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="65" data-offset-y="100" data-expected-width="135" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="40" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected 30 but got -225
FAIL .inline-grid 20 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline columns min-content-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="-10" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="100" data-expected-width="95" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected -10 but got 25
FAIL .inline-grid 21 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline columns max-content-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="-10" data-offset-y="0" data-expected-width="75" data-expected-height="100">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="100" data-expected-width="95" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="200"></div>
</div>
offsetLeft expected -10 but got 25
FAIL .inline-grid 22 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline auto-columns fit-content-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="-10" data-offset-y="0" data-expected-width="75" data-expected-height="50">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="50" data-expected-width="95" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="150"></div>
</div>
offsetLeft expected -10 but got 25
FAIL .inline-grid 23 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline auto-columns flex-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="-10" data-offset-y="0" data-expected-width="75" data-expected-height="50">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="50" data-expected-width="95" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="150"></div>
</div>
offsetLeft expected -10 but got 25
FAIL .inline-grid 24 assert_equals:
<div class="inline-grid verticalRL alignItemsBaseline auto-columns max-flex-rows">
  <div class="firstRowFirstColumn bigFont paddingLeft" data-offset-x="-10" data-offset-y="0" data-expected-width="75" data-expected-height="50">É</div>
  <div class="firstRowSecondColumn horizontalTB" data-offset-x="25" data-offset-y="50" data-expected-width="95" data-expected-height="100">ÉÉ É ÉÉ ÉÉÉÉ É ÉÉ ÉÉÉ ÉÉ É</div>
  <div class="autoRowAutoColumnSpanning2 width25" data-offset-x="0" data-offset-y="0" data-expected-width="25" data-expected-height="150"></div>
</div>
offsetLeft expected -10 but got 25
